Resolve namespace changes
