+++ R/strategy_support.R
@@ -51,16 +51,6 @@ make_environment <- function(type, ...) {
          stop("Unknown type ", type))
 }
 
-
-cohort_schedule_max_time_default <- function(p) {
-  cl <- class(p)[[1]]
-  switch(cl,
-         "Parameters<FF16,FF16_Env>"=`cohort_schedule_max_time_default__Parameters___FF16__FF16_Env`,
-         "Parameters<FF16r,FF16_Env>"=`cohort_schedule_max_time_default__Parameters___FF16r__FF16_Env`,
-         "Parameters<K93,K93_Env>"=`cohort_schedule_max_time_default__Parameters___K93__K93_Env`,
-         stop("Unknown type: ", cl))(p)
-}
-
 cohort_schedule_default <- function(p) {
   cl <- class(p)[[1]]
   switch(cl,
