+++ inst/include/plant/disturbance_regime.h
@@ -4,8 +4,6 @@
 
 #include <vector>
 
-using namespace Rcpp;
-
 namespace plant {
 
 class Disturbance_Regime {
@@ -14,15 +12,6 @@ public:
   double pr_survival(double time) const;
   std::vector<double> r_density(std::vector<double> time) const;
 };
-
-std::vector<double> Disturbance_Regime::r_density(std::vector<double> time) const {
-  std::vector<double> ret;
-  ret.reserve(time.size());
-  for (const auto& t : time) {
-    ret.push_back(density(t));
-  }
-  return ret;
-}
 
 }
