+++ inst/include/plant/stochastic_utils.h
@@ -3,7 +3,6 @@
 #define PLANT_PLANT_STOCHASTIC_UTILS_H_
 
 #include <plant/cohort_schedule.h>
-#include <plant/disturbance.h>
 #include <plant/scm_utils.h>
 
 namespace plant {
