+++ src/cohort_schedule.cpp
@@ -1,6 +1,5 @@
 #include <plant/cohort_schedule.h>
 #include <plant/parameters.h>
-#include <plant/disturbance.h>
 #include <plant/util.h>
 #include <Rcpp.h>
 #include <cmath> // log2, exp2
