+++ src/disturbance_regime.cpp
@@ -1,0 +1,14 @@
+#include <plant/disturbance_regime.h>
+
+namespace plant {
+
+std::vector<double> Disturbance_Regime::r_density(std::vector<double> time) const {
+  std::vector<double> ret;
+  ret.reserve(time.size());
+  for (const auto& t : time) {
+    ret.push_back(density(t));
+  }
+  return ret;
+}
+
+}
