+++ src/k93_cohort.cpp
@@ -4,11 +4,6 @@
 // Helpers for K93 model
 
 // [[Rcpp::export]]
-double cohort_schedule_max_time_default__Parameters___K93__K93_Env(const plant::Parameters<plant::K93_Strategy,plant::K93_Environment>& p) {
-   return plant::cohort_schedule_max_time_default<plant::Parameters<plant::K93_Strategy,plant::K93_Environment> >(p);
-}
-
-// [[Rcpp::export]]
 plant::CohortSchedule cohort_schedule_default__Parameters___K93__K93_Env(const plant::Parameters<plant::K93_Strategy,plant::K93_Environment>& p) {
    return plant::cohort_schedule_default<plant::Parameters<plant::K93_Strategy,plant::K93_Environment> >(p);
 }
